# packages/api-contract/openapi.yaml
openapi: 3.0.3
info:
  title: Mail Tracker API
  version: 1.0.0
servers:
  - url: /api/v1
security:
  - owner: []
paths:
...
  /documents/{documentId}:
    delete:
      operationId: deleteDocument
      parameters:
        - { name: documentId, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Shares revoked and private object deletion queued
          content:
            application/json:
              schema:
                { type: object, required: [deleted], properties: { deleted: { type: boolean } } }
  /documents:
    get:
      operationId: listDocuments
      responses:
        '200':
          description: Document operation result
          content:
            application/json:
              schema:
                { 'type': 'array', 'items': { '$ref': '#/components/schemas/HostedDocument' } }
    post:
      operationId: uploadDocument
      parameters:
        - {
            name: X-File-Name,
            in: header,
            required: true,
            schema: { type: string },
            description: Percent-encoded UTF-8 filename,
          }
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema: { type: string, format: binary }
      responses:
        '200':
          description: Private immutable document uploaded using Raildrop storage
          content:
            application/json:
              schema: { '$ref': '#/components/schemas/HostedDocument' }
  /documents/{documentId}/shares:
    get:
      operationId: listDocumentShares
      parameters:
        - { name: documentId, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Document operation result
          content:
            application/json:
              schema: { 'type': 'array', 'items': { '$ref': '#/components/schemas/DocumentShare' } }
    post:
      operationId: createDocumentShare
      parameters:
        - { name: documentId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              {
                type: object,
                required: [days, allowDownload],
                properties:
                  {
                    days: { type: integer, minimum: 1, maximum: 365 },
                    allowDownload: { type: boolean },
                  },
              }
      responses:
        '200':
          description: Expiring document access link
          content:
            application/json:
              schema: { '$ref': '#/components/schemas/DocumentShare' }
  /document-shares/{shareId}/revoke:
    post:
      operationId: revokeDocumentShare
      parameters:
        - { name: shareId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        '200':
          description: Document operation result
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
  /viewer/sessions:
    post:
      operationId: startViewerSession
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              {
                'type': 'object',
                'required': ['token'],
                'properties': { 'token': { 'type': 'string' } },
              }
      responses:
        '200':
          description: Document operation result
          content:
            application/json:
              schema: { '$ref': '#/components/schemas/ViewerSession' }
  /viewer/{sessionId}:
    get:
      operationId: getViewerSession
      security: []
      parameters:
        - { name: sessionId, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Document operation result
          content:
            application/json:
              schema: { '$ref': '#/components/schemas/ViewerSession' }
  /viewer/{sessionId}/events:
    post:
      operationId: recordViewerEvents
      security: []
      parameters:
        - { name: sessionId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { '$ref': '#/components/schemas/ViewerBatch' }
      responses:
        '200':
          description: Document operation result
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
  /viewer/{sessionId}/file:
    get:
      operationId: getViewerFile
      security: []
      parameters:
        - { name: sessionId, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Session-authorized binary file; byte ranges supported
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
  /viewer/{sessionId}/download:
    get:
      operationId: downloadViewerFile
      security: []
      parameters:
        - { name: sessionId, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Session-authorized binary file; byte ranges supported
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
...
components:
  securitySchemes:
    owner: { type: http, scheme: bearer, bearerFormat: JWT }
  schemas:
...
    HostedDocument:
      {
        'type': 'object',
        'required':
          [
            'id',
            'filename',
            'contentType',
            'sizeBytes',
            'createdAt',
            'sessions',
            'activeSeconds',
            'downloads',
          ],
        'properties':
          {
            'id': { 'type': 'string' },
            'filename': { 'type': 'string' },
            'contentType':
              { 'type': 'string', 'enum': ['application/pdf', 'image/jpeg', 'image/png'] },
            'sizeBytes': { 'type': 'integer' },
            'createdAt': { 'type': 'string', 'format': 'date-time' },
            'sessions': { 'type': 'integer' },
            'activeSeconds': { 'type': 'number' },
            'downloads': { 'type': 'integer' },
          },
      }
    DocumentShare:
      {
        'type': 'object',
        'required': ['id', 'documentId', 'expiresAt', 'revokedAt', 'allowDownload'],
        'properties':
          {
            'id': { 'type': 'string' },
            'documentId': { 'type': 'string' },
            'url': { 'type': 'string' },
            'expiresAt': { 'type': 'string', 'format': 'date-time' },
            'revokedAt': { 'type': 'string', 'format': 'date-time', 'nullable': true },
            'allowDownload': { 'type': 'boolean' },
          },
      }
    ViewerSession:
      {
        'type': 'object',
        'required':
          [
            'id',
            'documentId',
            'filename',
            'contentType',
            'sizeBytes',
            'allowDownload',
            'expiresAt',
          ],
        'properties':
          {
            'id': { 'type': 'string' },
            'documentId': { 'type': 'string' },
            'filename': { 'type': 'string' },
            'contentType': { 'type': 'string' },
            'sizeBytes': { 'type': 'integer' },
            'allowDownload': { 'type': 'boolean' },
            'expiresAt': { 'type': 'string', 'format': 'date-time' },
          },
      }
    ViewerBatch:
      {
        'type': 'object',
        'required': ['batchId', 'kind'],
        'properties':
          {
            'batchId': { 'type': 'string', 'format': 'uuid' },
            'kind':
              { 'type': 'string', 'enum': ['viewer_loaded', 'viewer_activity', 'page_exposure'] },
            'activeMs': { 'type': 'integer', 'minimum': 0, 'maximum': 15000 },
            'pages':
              {
                'type': 'array',
                'maxItems': 100,
                'items': { 'type': 'integer', 'minimum': 1, 'maximum': 1000 },
              },
          },
      }
...